Move reporting stuff to report.rs from lib.rs
